requirements.txt
pyyaml
kubernetes
matplotlib